registration rework, new generic class for trains and rollingstock
- Block and item registration for normal things is now done from the
same class.
- Train item registration is now handled through thr train registry like
the rest of the train stuff.
- The train item has been vastly simplified by offloading the majority
of the code that would be generic to all trains and rollingstock to the
RailUtility.
- Created a new class to house the functionality that is the same for
trains and rollingstock, trains and rollingstock will still have their
own classes to extend this new class, but overall this helps to make
things more generic and organized.
- Some documentation was reworked, mostly to better explain the process
of adding new trains.