Smoke, steam and bogies
- Trains now properly display smoke, steam, etc. The system for it has
been fully reworked and requires a little more effort to use, but it
adds a TON of features to make up for it, everything from smoke color to
density, multiple points per train, and custom physics for the effects
that are not only lighter than minecraft's default, but more realistic.
- Added a baseline system to render bogies, for when we actually have
bogies.