train creative mode, GUI and fuel reworks, and more
- Trains now have a creative mode, when this is enabled they will not
remove the fuel item when filling the fuel. for example an electric
train will have infinate power from a single redstone dust, or a steam
train infinate water from a single bucket (there is still no GUI button
for this yet, but soon).
- Added a variable to track the linked trains/rollingstock, not used
yet.
- The GUI has undergone a massive redesign adding in hover text for fuel
tanks, placement for the soon to be added train toolbar, repositioning
to get ready for the GUI graphics, and better space allocation to allow
for larger train inventories (up to and maybe past 6x6).
- The animations are now linked with the train's current speed.
- The values for steam use have been re-calculated, it may need some
more blalacing later but it's actually functional now and at least
partially realistic.
- Fuel management for electric trains has been added.