GUI progress, NBT and fixes.
- Fixed a mistake in the readme, the scale was significantly incorrect.
- Added some variables to NBT so they actually save.
- Added the train's toolbar, this will hold a bunch of toggles to
control different features of the train, they are treated like item
slots to better fit with the UI.
- Fixed some null pointer issues in the inventory handler.
- Reworked the GUI to reuse elements from the game's inbuilt UI. The
upper background, and some other aspects will still have their own
graphics later on but now the slots and player inventory will go with
the current resourcepack, even if it wasint intended to support this
mod.
- Added in the networking for the train's toggles.
- Added a basic button that I can later use to make the rest of the
GUI's buttons.