Documentation, fuel, crafting, and more reworks.
- Names have been reworked, now the item name will also define the name
of the entity, it's inventory, the item's texture, and the name that
shows in the GUI. Two less fields needed to make a new entity.
- Documentation reworks, I didn't have time to finish it but I reworked
a small portion of the documentation and added some new documentation.
- Removed two unnecessary classes.
- Renamed the train class to match what train it is.
- Crafting has been reworked and cleaned up, functionality for crafting
from tile entities/blocks have been expanded, and rollingstock with
inbuilt workbenches will be easier to implement in the future.
- Inventory code has been reworked to work better alongside the crafting
functionality.
- Items will now have description text to show more information for the
train/rollingstock.
- Added some entries to the URI registry for item textures.
- Moved the storage entity to the tile entities folder, we no longer
need a crafting folder.
- Added a null value for inventory sizes.
- Optimized the imports.