motion update and bugfix
- TESTING CHANGE: pushing a rollingstock now makes it go a lot faster,
and they increase speed with time.
- Progress towards new motion, it's no more or less reliable than
vanilla, but it should be... its close to what is needed, but not there
yet.
- Entity destuction changes, moved it from the hitbox to the core
entity, now the entity can drop an item of itself on death, it also can
be damaged by normal means, and has health so a single strike won't
always brake it.
- Transports can now be reinforced, defined in the transport class, this
allows it to take less damage.
- You can now define the render scale of the transport, good for if you
made it too wide to fit the rails.
- Fixed an improper calculation in the rail util, mostly.