massive reworks
- Reworked the CMD Bogie model.
- Reworked the GATX 13000 Gallon Tanker model.
- Bogies, Seats, and hitbox should no longer be persistent.
- Fixed seat rider position.
- Fixed trains moving without fuel, and not capping max speed.
- Fixed some inventory bugs.
- Fixed inability to remove a transport from world when the hitboxes
don't spawn.
- Fixed some positions in the GATX 13000 Gallon Tanker.
- Updated fuel management, now it can be overwridden in the API, and can
manage things like consuming fluids from buckets for tanker cars.
- Added a custom button class for buttons with hover text rather than
buttons with text in them, it can be expanded later to also handle the
icons as well, to significantly simplify the GUI code.
- Full GUI rework, simplified code and management, more space in larger
cars, support for inventories up to 9x12, and better overall
organization.
- New animations for GroupedModelRenderer, used to scale geometry based
on inventory useage percentage, with either the default texture of the
model or the texture of the relevant block/item.
- Reworked the particle render, now properly instances the tessellator,
because apparently that's important.
- Fixed a glitch in the tessellator due to the GL calls not being
cleared when they are supposed to be.
- Added a backup function for client to send a packet to server for
forcing respawn of hitboxes and bogies.
- Updated hitboxes, now they properly push entities, and eachother,
away.
- Added an experimental ore directory manager to the RailUtility.
- Added tanker support to the slot manager.
- Resized the base textures to their proper scale.