rail improvements
So TL;DR, tried to implement rails that can be submerged, but thats not possible because water blocks get the material from a variable that's write protected in any block that extends vanilla rails. so that was a LOT of time wasted, all the related functionality has subsequently been removed.
ANYWAY
- Rail blocks can now have their gauge and model scale defined on registration, gauge defines how far apart the rails are, scale defines how big the model actually renders.
- Registration will now actually prevent blocks from registering with the same names. this also applies to tile entities to prevent different blocks from registering the same tile entity multiple times.
- Fixed a registration bug related to buckets.
- Added a debug option to intentionally crash the game with a stacktrace, only wokrs when the game is being run from an IDE.
- Removed a large number of unnecessary imports.